typechange

+++ interface/GetNames.h
@@ -12,7 +12,7 @@ std::string GetCalibrationFilename(int telescopeID);
 std::string GetAlignmentFilename(int telescopeID, bool useInitial=0);
 std::string GetMaskingFilename(int telescopeID);
 const char * GetSignalBranchName();
-int GetNumberOfROCS(int16_t);
+uint8_t GetNumberOfROCS(int16_t);
 int GetNumberOfSignals(int16_t);
 int GetUseGainInterpolator(int telescopeID);
 int GetUseExternalCalibrationFunction(int telescopeID);
